extract Speaker to separate component in legacy draft
